moved slice visualizer to separate folder, slicer table wip
