+++ VSRAD.Package/DebugVisualizer/SliceVisualizer/SliceVisualizerControl.xaml
@@ -1,11 +1,10 @@
-﻿<UserControl x:Class="VSRAD.Package.DebugVisualizer.SliceVisualizerControl"
+﻿<UserControl x:Class="VSRAD.Package.DebugVisualizer.SliceVisualizer.SliceVisualizerControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:VSRAD.Package.DebugVisualizer"
              mc:Ignorable="d">
-    <!--d:DesignHeight="450" d:DesignWidth="800"-->
+            <!-- DesignHeight="450" d:DesignWidth="800"> -->
     <DockPanel>
         <TextBlock Text="Watch: "/>
         <ComboBox ItemsSource="{Binding Watches}"/>
