+++ VSRAD.Package/DebugVisualizer/SliceVisualizer/SliceVisualizerControl.xaml.cs
@@ -1,17 +1,31 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
 using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
 using VSRAD.Package.Options;
 using VSRAD.Package.ToolWindows;
 using VSRAD.Package.Utils;
 
-namespace VSRAD.Package.DebugVisualizer
+namespace VSRAD.Package.DebugVisualizer.SliceVisualizer
 {
     /// <summary>
     /// Interaction logic for SliceVisualizerControl.xaml
     /// </summary>
     public partial class SliceVisualizerControl : UserControl
     {
+
+        private Server.BreakState _breakState;
+
         public sealed class Context : DefaultNotifyPropertyChanged
         {
             public ProjectOptions Options { get; }
@@ -25,7 +39,7 @@ namespace VSRAD.Package.DebugVisualizer
                 Options = options;
             }
         }
-        
+
 
         public SliceVisualizerControl(IToolWindowIntegration integration)
         {
